+++ src/main/kotlin/com/github/jcornaz/kable/Table.kt
@@ -39,10 +39,6 @@ interface Table<R, C, out V> {
     /** Entries of the table */
     val entries: Set<Entry<R, C, V>>
 
-    /** Return a map representation of the table */
-    fun toMap(): Map<Pair<R, C>, V> =
-            entries.associate { it.toPair() }
-
     /** Return true if, and only if, the table has no entry */
     fun isEmpty(): Boolean = size == 0