+++ src/main/kotlin/com/github/jcornaz/kable/impl/EmptyTable.kt
@@ -33,8 +33,6 @@ object EmptyTable : AbstractTable<Any?, Any?, Any?>() {
     override val values = emptyList<Nothing>()
     override val entries = emptySet<Entry<Any?, Any?, Any?>>()
 
-    override fun toMap() = emptyMap<Pair<Any?, Any?>, Any?>()
-
     override fun isEmpty() = true
 
     override fun containsRow(row: Any?) = false