+++ src/main/kotlin/com/github/jcornaz/kable/util/Collections.kt
@@ -40,8 +40,14 @@ fun <R, C, V> Iterable<Entry<R, C, V>>.toTable(): Table<R, C, V> {
     return BiKeyMap(this)
 }
 
+/** Create a new table with this entries */
+fun <R, C, V> Sequence<Entry<R, C, V>>.toTable(): Table<R, C, V> = asIterable().toTable()
+
 /** Create a new mutable table with this entries */
 fun <R, C, V> Iterable<Entry<R, C, V>>.toMutableTable(): MutableTable<R, C, V> = ThreeMapTable(this)
+
+/** Create a new mutable table with this entries */
+fun <R, C, V> Sequence<Entry<R, C, V>>.toMutableTable(): MutableTable<R, C, V> = asIterable().toMutableTable()
 
 /**
  * Groups elements of the original collection by the key returned by the given [keySelector] function
@@ -50,7 +56,7 @@ fun <R, C, V> Iterable<Entry<R, C, V>>.toMutableTable(): MutableTable<R, C, V> =
  * The returned table preserves the entry iteration order of the keys produced from the original collection.
  */
 inline fun <R, C, T> Iterable<T>.groupTableBy(keySelector: (T) -> Pair<R, C>): Table<R, C, List<T>> =
-        groupBy(keySelector).toTable()
+        asSequence().groupBy(keySelector).toTable()
 
 /**
  * Groups values returned by the [valueTransform] function applied to each element of the original collection
@@ -60,7 +66,7 @@ inline fun <R, C, T> Iterable<T>.groupTableBy(keySelector: (T) -> Pair<R, C>): T
  * The returned table preserves the entry iteration order of the keys produced from the original collection.
  */
 inline fun <R, C, V, T> Iterable<T>.groupTableBy(keySelector: (T) -> Pair<R, C>, valueTransform: (T) -> V): Table<R, C, List<V>> =
-        groupBy(keySelector, valueTransform).toTable()
+        asSequence().groupBy(keySelector, valueTransform).toTable()
 
 /**
  * Returns a [Table] containing key-value pairs provided by [transform] function
@@ -70,8 +76,8 @@ inline fun <R, C, V, T> Iterable<T>.groupTableBy(keySelector: (T) -> Pair<R, C>,
  *
  * The returned table preserves the entry iteration order of the original collection.
  */
-inline fun <R, C, V, T> Iterable<T>.associateTable(transform: (T) -> Entry<R, C, V>): Table<R, C, V> =
-        map(transform).toTable()
+fun <R, C, V, T> Iterable<T>.associateTable(transform: (T) -> Entry<R, C, V>): Table<R, C, V> =
+        asSequence().map(transform).toTable()
 
 /**
  * Returns a [Table] containing the elements from the given collection indexed by the key
@@ -80,7 +86,7 @@ inline fun <R, C, V, T> Iterable<T>.associateTable(transform: (T) -> Entry<R, C,
  * If any two elements would have the same key returned by [keySelector] the last one gets added to the table.
  */
 inline fun <R, C, V> Iterable<V>.associateTableBy(keySelector: (V) -> Pair<R, C>): Table<R, C, V> =
-        associateBy(keySelector).toTable()
+        asSequence().associateBy(keySelector).toTable()
 
 /**
  * Returns a [Table] containing the values provided by [valueTransform] and indexed by [keySelector] functions applied to elements of the given collection.
@@ -88,4 +94,4 @@ inline fun <R, C, V> Iterable<V>.associateTableBy(keySelector: (V) -> Pair<R, C>
  * If any two elements would have the same key returned by [keySelector] the last one gets added to the table.
  */
 inline fun <R, C, V, T> Iterable<T>.associateTableBy(keySelector: (T) -> Pair<R, C>, valueTransform: (T) -> V): Table<R, C, V> =
-        associateBy(keySelector, valueTransform).toTable()
+        asSequence().associateBy(keySelector, valueTransform).toTable()