+++ src/main/kotlin/com/github/jcornaz/kable/util/Tables.kt
@@ -61,6 +61,17 @@ fun <R, C, V> Table<R, C, V>.toMutableTable(): MutableTable<R, C, V> = entries.t
 /** Return true if, and only if, the table contains the given row-column key pair */
 operator fun <R, C> Table<R, C, *>.contains(key: Pair<R, C>) = contains(key.first, key.second)
 
+/** Return a map representation of the table */
+fun <R, C, V> Table<R, C, V>.toMap(): Map<Pair<R, C>, V> = entries.associate { it.toPair() }
+
+/** Create a sequence that returns all entries of this table */
+fun <R, C, V> Table<R, C, V>.asSequence(): Sequence<Entry<R, C, V>> = asIterable().asSequence()
+
+/** Create an iterable that returns all entries of this table */
+fun <R, C, V> Table<R, C, V>.asIterable(): Iterable<Entry<R, C, V>> = object : Iterable<Entry<R, C, V>> {
+    override fun iterator() = this@asIterable.iterator()
+}
+
 /**
  * Return all the values of the row mapped by column
  *
@@ -70,116 +81,119 @@ operator fun <R, C, V> Table<R, C, V>.get(row: R): Map<C, V> = getRow(row)
 
 /** Return a new table with the same entries plus the given new entry */
 operator fun <R, C, V> Table<R, C, V>.plus(entry: Entry<R, C, V>): Table<R, C, V> =
-        (toMap() + entry.toPair()).toTable()
+        (asSequence() + entry).toTable()
 
 /** Return a new table with the same entries plus the given new entries */
 operator fun <R, C, V> Table<R, C, V>.plus(entries: Collection<Entry<R, C, V>>): Table<R, C, V> =
-        (toMap() + entries.associate { it.toPair() }).toTable()
+        (asSequence() + entries).toTable()
 
 /** Return a new table with the same entries plus entries of the given table */
 operator fun <R, C, V> Table<R, C, V>.plus(table: Table<R, C, V>): Table<R, C, V> =
-        (toMap() + table.toMap()).toTable()
+        (asSequence() + table.asSequence()).toTable()
 
 /** Return a new table with the same entries except the given row */
 fun <R, C, V> Table<R, C, V>.minusRow(row: R): Table<R, C, V> =
         if (!containsRow(row)) this
-        else toMap().filterNot { (key, _) -> key.first == row }.toTable()
+        else asSequence().filterNot { it.row == row }.toTable()
 
 /** Return a new table with the same entries except the given column */
 fun <R, C, V> Table<R, C, V>.minusColumn(column: C): Table<R, C, V> =
         if (!containsColumn(column)) this
-        else toMap().filterNot { (key, _) -> key.second == column }.toTable()
+        else asSequence().filterNot { it.column == column }.toTable()
 
 /** Return a new table with the same entries except the row-column pair */
 operator fun <R, C, V> Table<R, C, V>.minus(key: Pair<R, C>): Table<R, C, V> =
         if (key !in this) this
-        else toMap().filterNot { (k, _) -> k.first == key.first && k.second == key.second }.toTable()
+        else asSequence().filterNot { it.row == key.first && it.column == key.second }.toTable()
 
 /** Performs the given [action] on each entry */
 inline fun <R, C, V> Table<R, C, V>.forEach(action: (Entry<R, C, V>) -> Unit) =
-        entries.forEach(action)
+        asSequence().forEach(action)
 
 /** Returns `true` if all entries match the given [predicate] */
 inline fun <R, C, V> Table<R, C, V>.all(predicate: (Entry<R, C, V>) -> Boolean): Boolean =
-        toMap().all { predicate(it.toTableEntry()) }
+        asSequence().all(predicate)
 
 /** Returns `true` if map has at least one entry */
-fun Table<*, *, *>.any(): Boolean = toMap().any()
+fun Table<*, *, *>.any(): Boolean =
+        asSequence().any()
 
 /** Returns `true` if at least one entry matches the given [predicate] */
 inline fun <R, C, V> Table<R, C, V>.any(predicate: (Entry<R, C, V>) -> Boolean): Boolean =
-        toMap().any { predicate(it.toTableEntry()) }
+        asSequence().any(predicate)
 
 /** Returns `true` if the map has no entries */
-fun Table<*, *, *>.none(): Boolean = toMap().none()
+fun Table<*, *, *>.none(): Boolean =
+        asSequence().none()
 
 /** Returns `true` if no entry match the given [predicate] */
 inline fun <R, C, V> Table<R, C, V>.none(predicate: (Entry<R, C, V>) -> Boolean): Boolean =
-        toMap().none { predicate(it.toTableEntry()) }
+        asSequence().none(predicate)
 
 /** Returns the number of entries in this map */
 fun Table<*, *, *>.count(): Int = size
 
 /** Returns the number of entries matching the given [predicate] */
 inline fun <R, C, V> Table<R, C, V>.count(predicate: (Entry<R, C, V>) -> Boolean): Int =
-        toMap().count { predicate(it.toTableEntry()) }
+        asSequence().count(predicate)
 
 /** Returns the first entry yielding the largest value of the given function or `null` if there are no entries */
 inline fun <R, C, V, T : Comparable<T>> Table<R, C, V>.maxBy(selector: (Entry<R, C, V>) -> T): Entry<R, C, V>? =
-        entries.maxBy(selector)
+        asSequence().maxBy(selector)
 
 /** Returns the first entry having the largest value according to the provided [comparator] or `null` if there are no entries */
 fun <R, C, V> Table<R, C, V>.maxWith(comparator: Comparator<Entry<R, C, V>>): Entry<R, C, V>? =
-        entries.maxWith(comparator)
+        asSequence().maxWith(comparator)
 
 /** Returns the first entry yielding the smallest value of the given function or `null` if there are no entries */
 inline fun <R, C, V, T : Comparable<T>> Table<R, C, V>.minBy(selector: (Entry<R, C, V>) -> T): Entry<R, C, V>? =
-        entries.minBy(selector)
+        asSequence().minBy(selector)
 
 /** Returns the first entry having the smallest value according to the provided [comparator] or `null` if there are no entries */
 fun <R, C, V> Table<R, C, V>.minWith(comparator: Comparator<Entry<R, C, V>>): Entry<R, C, V>? =
-        entries.minWith(comparator)
+        asSequence().minWith(comparator)
 
 /** Returns a new table containing all entries matching the given [predicate] */
-inline fun <R, C, V> Table<R, C, V>.filter(predicate: (Entry<R, C, V>) -> Boolean): Table<R, C, V> =
-        toMap().filter { predicate(it.toTableEntry()) }.toTable()
+fun <R, C, V> Table<R, C, V>.filter(predicate: (Entry<R, C, V>) -> Boolean): Table<R, C, V> =
+        asSequence().filter(predicate).toTable()
 
 /** Returns a new table containing all entries not matching the given [predicate] */
-inline fun <R, C, V> Table<R, C, V>.filterNot(predicate: (Entry<R, C, V>) -> Boolean): Table<R, C, V> =
-        toMap().filterNot { predicate(it.toTableEntry()) }.toTable()
+fun <R, C, V> Table<R, C, V>.filterNot(predicate: (Entry<R, C, V>) -> Boolean): Table<R, C, V> =
+        asSequence().filterNot(predicate).toTable()
 
 /** Returns a new table containing all entries on rows matching the given [predicate] */
-inline fun <R, C, V> Table<R, C, V>.filterRows(predicate: (R) -> Boolean): Table<R, C, V> =
-        toMap().filterKeys { predicate(it.first) }.toTable()
+fun <R, C, V> Table<R, C, V>.filterRows(predicate: (R) -> Boolean): Table<R, C, V> =
+        asSequence().filter { predicate(it.row) }.toTable()
 
 /** Returns a new table containing all entries on columns matching the given [predicate] */
-inline fun <R, C, V> Table<R, C, V>.filterColumns(predicate: (C) -> Boolean): Table<R, C, V> =
-        toMap().filterKeys { predicate(it.second) }.toTable()
+fun <R, C, V> Table<R, C, V>.filterColumns(predicate: (C) -> Boolean): Table<R, C, V> =
+        asSequence().filter { predicate(it.column) }.toTable()
 
 /** Returns a new table containing all entries with a values matching the given [predicate] */
-inline fun <R, C, V> Table<R, C, V>.filterValues(predicate: (V) -> Boolean): Table<R, C, V> =
-        toMap().filterValues(predicate).toTable()
+fun <R, C, V> Table<R, C, V>.filterValues(predicate: (V) -> Boolean): Table<R, C, V> =
+        asSequence().filter { predicate(it.value) }.toTable()
 
 /** Returns a list containing the results of applying the given [transform] function to each entry in the original table */
-inline fun <R, C, V, T> Table<R, C, V>.map(transform: (Entry<R, C, V>) -> T): Collection<T> =
-        toMap().map { transform(it.toTableEntry()) }
+fun <R, C, V, T> Table<R, C, V>.map(transform: (Entry<R, C, V>) -> T): Collection<T> =
+        asSequence().map(transform).toList()
 
 /** Returns a list containing the results of applying the given [transform] function to each entry in the original table */
-inline fun <R, C, V, T> Table<R, C, V>.flatMap(transform: (Entry<R, C, V>) -> Iterable<T>): Collection<T> =
-        toMap().flatMap { transform(it.toTableEntry()) }
+fun <R, C, V, T> Table<R, C, V>.flatMap(transform: (Entry<R, C, V>) -> Iterable<T>): Collection<T> =
+        asIterable().flatMap(transform)
 
 /** Returns a list containing only the non-null results of applying the given [transform] function to each entry in the original table */
-inline fun <R, C, V, T : Any> Table<R, C, V>.mapNotNull(transform: (Entry<R, C, V>) -> T?): Collection<T> =
-        toMap().mapNotNull { transform(it.toTableEntry()) }
+fun <R, C, V, T : Any> Table<R, C, V>.mapNotNull(transform: (Entry<R, C, V>) -> T?): Collection<T> =
+        asSequence().mapNotNull(transform).toList()
 
 /** Returns a new table with entries having the rows obtained by applying the [transform] function to each entry */
-inline fun <R, C, V, T> Table<R, C, V>.mapRows(transform: (Entry<R, C, V>) -> T): Table<T, C, V> =
-        toMap().mapKeys { transform(it.toTableEntry()) to it.key.second }.toTable()
+fun <R, C, V, T> Table<R, C, V>.mapRows(transform: (Entry<R, C, V>) -> T): Table<T, C, V> =
+        asSequence().map { entry(transform(it), it.column, it.value) }.toTable()
 
 /** Returns a new table with entries having the columns obtained by applying the [transform] function to each entry */
-inline fun <R, C, V, T> Table<R, C, V>.mapColumns(transform: (Entry<R, C, V>) -> T): Table<R, T, V> =
-        toMap().mapKeys { it.key.first to transform(it.toTableEntry()) }.toTable()
+fun <R, C, V, T> Table<R, C, V>.mapColumns(transform: (Entry<R, C, V>) -> T): Table<R, T, V> =
+        asSequence().map { entry(it.row, transform(it), it.value) }.toTable()
 
 /** Returns a new table with entries having the values obtained by applying the [transform] function to each entry */
-inline fun <R, C, V, T> Table<R, C, V>.mapValues(transform: (Entry<R, C, V>) -> T): Table<R, C, T> =
-        toMap().mapValues { transform(it.toTableEntry()) }.toTable()
+fun <R, C, V, T> Table<R, C, V>.mapValues(transform: (Entry<R, C, V>) -> T): Table<R, C, T> =
+        asSequence().map { entry(it.row, it.column, transform(it)) }.toTable()
+